fix: actualQuantity/expectedQuantity dùng BigDecimal thay Integer
InventoryBalanceDTO và InventoryBalanceViewDTO dùng Integer cho
quantity → truncate phần thập phân (154.9 → 154). Đổi sang BigDecimal
để giữ chính xác cho sản phẩm đơn vị mét.

+++ src/main/java/manage/store/inventory/dto/InventoryBalanceDTO.java
@@ -1,5 +1,7 @@
 package manage.store.inventory.dto;
 
+import java.math.BigDecimal;
+
 public interface InventoryBalanceDTO {
 
     Long getVariantId();
@@ -10,6 +12,6 @@ public interface InventoryBalanceDTO {
     String getItemCode();
     String getItemName();
     String getUnit();
-    Integer getActualQuantity();
-    Integer getExpectedQuantity();
+    BigDecimal getActualQuantity();
+    BigDecimal getExpectedQuantity();
 }
